Make token validation constant-time
the `check_header_password()` function is doing byte-at-a-time equality checking to check if the token is valid(i.e. submitted_token == valid_token).  This method leaks timing information over the network, and can be used by an attacker to mount a timing attack against the token.

Recommend the use of `hmac.compare_digest` to mitigate against the possibility of timing attacks against the token.